relise du programe à l'état initial

+++ TP1/prog1.py
@@ -4,9 +4,9 @@ import fonctions as f
 
 def main():
     try:
-        # Tester avec des flottants au lieu des entiers
-        a = float(input("Entrez la base (a) : "))  # Passez un flottant pour le test
-        b = float(input("Entrez l'exposant (b) : "))  # Passez un flottant pour le test
+        # Saisie des nombres
+        a = int(input("Entrez la base (a) : "))
+        b = int(input("Entrez l'exposant (b) : "))
         
         # Appel de la fonction puissance et affichage du résultat
         res = f.puissance(a, b)
